+++ src/test/java/ru/r2cloud/sdrmodem/SdrModemClientTest.java
@@ -111,7 +111,7 @@ public class SdrModemClientTest {
 		File wav = TestUtil.setupClasspathResource(tempFolder, "data/delfipq.raw.gz");
 		Decoder decoder = decoders.findByTransmitter(transmitter);
 		DecoderResult result = decoder.decode(wav, req, transmitter);
-		assertEquals(1, result.getNumberOfDecodedPackets().longValue());
+		assertEquals(1, result.getNumberOfDecodedPackets());
 		try (BeaconInputStream<DelfiPqBeacon> source = new BeaconInputStream<>(new BufferedInputStream(new FileInputStream(result.getDataPath())), DelfiPqBeacon.class)) {
 			assertTrue(source.hasNext());
 			AssertJson.assertObjectsEqual("DelfiPqBeacon.json", source.next());
